refactor : course에 강의리스트 초기값 세팅 추가
