+++ src/main/java/nextstep/courses/domain/Course.java
@@ -22,6 +22,9 @@ public class Course {
     public Course(String title, Long creatorId) {
         this(0L, title, creatorId, LocalDateTime.now(), null,1);
     }
+    public Course(Long id, String title, Long creatorId, List<Session> sessions) {
+        this(id, title, creatorId, LocalDateTime.now(), null,1, sessions);
+    }
 
     public Course(Long id, String title, Long creatorId, LocalDateTime createdAt, LocalDateTime updatedAt, int classNumber) {
         this(id, title, creatorId, createdAt, updatedAt, classNumber, new ArrayList<>());
@@ -33,8 +36,8 @@ public class Course {
         this.creatorId = creatorId;
         this.createdAt = createdAt;
         this.updatedAt = updatedAt;
-        this.sessions = sessions;
         this.classNumber = classNumber;
+        this.sessions = sessions;
     }
 
     public String getTitle() {
@@ -57,12 +60,12 @@ public class Course {
     public String toString() {
         return "Course{" +
                 "id=" + id +
+                ", classNumber=" + classNumber +
                 ", title='" + title + '\'' +
                 ", creatorId=" + creatorId +
                 ", createdAt=" + createdAt +
                 ", updatedAt=" + updatedAt +
+                ", sessions=" + sessions +
                 '}';
     }
-
-
 }
